+++ packages/components/mapbox/src/index.js
@@ -1,4 +1,3 @@
 export {default as CustomControl} from './CustomControl.svelte';
 export {default as Mapbox} from './Mapbox.svelte';
 export {default as SvgLayer} from './SvgLayer.svelte';
-export * from './consts.js';