+++ packages/docs/site/src/routes/compounds/mapbox/custom_controls.svelte
@@ -2,7 +2,7 @@
 	import {CustomControl, Mapbox} from '@svizzle/mapbox';
 
 	const accessToken = 'pk.eyJ1IjoibmVzdGEtdWsiLCJhIjoiY2ozbjUzY2drMDAwNzJxbnl6a21uM253cSJ9.3RTMySEVk0LC4gQvGoG-Zw';
-	const styleURL = 'mapbox://styles/nesta-uk/cl8olrzo200ci16pim0h4c1pn';
+	const style = 'mapbox://styles/nesta-uk/cl8olrzo200ci16pim0h4c1pn';
 
 	let bounds;
 	const zoomToEngland = () => {
@@ -14,7 +14,7 @@
 	<Mapbox
 		{accessToken}
 		{bounds}
-		{styleURL}
+		{style}
 		withScaleControl={true}
 		withZoomControl={true}
 	>